requirements.txt
tensorflow>=2.13.0
keras>=2.13.0
streamlit>=1.32.0
numpy>=1.24.0
Pillow>=9.4.0
matplotlib>=3.7.0
tqdm>=4.65.0
nltk>=3.8.0          # optional: BLEU evaluation
